+++ commons/api/prestation.py
@@ -1,21 +1,21 @@
 from tastypie.resources import ModelResource
 from tastypie.authorization import Authorization
 from tastypie import fields
-from .models import Prestation, PrestationModule, SelectedModules
+from commons.models.prestation import Prestation, PrestationModule, SelectedModules
 
 
 class PrestationModuleResource(ModelResource):
     class Meta:
         queryset = PrestationModule.objects.all()
         allowed_methods = ['get']
-        resource_name = 'prestation/modules'
+        resource_name = 'project/commons/prestationmodules'
         authorization = Authorization()
 
 class PrestationResource(ModelResource):
     class Meta:
         queryset = Prestation.objects.all()
         allowed_methods = ['get']
-        resource_name = 'prestation/prestation'
+        resource_name = 'project/commons/prestation'
         authorization = Authorization()
 
     modules = fields.ToManyField(PrestationModuleResource, 'module', null=True, blank=True, full=True)
