+++ commons/models/__init__.py
@@ -1,0 +1,3 @@
+# __init__.py
+from prestation import PrestationModule, Prestation, SelectedModules
+from usage import Usage, Pertinence
