+++ dataserver/settings.py
@@ -163,7 +163,6 @@ INSTALLED_APPS = (
     'scout',
     'transport_vlille',
     'unisson',
-    'prestation'
 )
 
 # A sample logging configuration. The only tangible logging
